+++ assets/javascripts/kitten/components/form/checkbox/index.js
@@ -45,7 +45,7 @@ const StyledCheckbox = styled.div`
     margin: 0 ${pxToRem(10)} 0 0;
 
     background-color: ${COLORS.background1};
-    border: ${pxToRem(1)} solid var(--color-grey-400);
+    border: var(--border-width-1) solid var(--color-grey-400);
     cursor: pointer;
     transition: border-color 0.2s, background-color 0.2s;