+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Design: Change `borderWidth` in `1px` and used A11Y color standards.
+- Token: Add `--border-width-1` variable.
+- Feature: Design: Used `--border-width-1` and used A11Y color standards.
 - Feature: `DashboardMenu`: Add `size` prop to `DashboardMenu.Expandable`, defaults to `default`.
 - Fix: `DashboardMenu`: Style fixes.